requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
attrs==24.2.0
azure-ai-ml==1.20.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.18.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.23.1
azure-storage-file-datalake==12.17.0
azure-storage-file-share==12.18.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
filelock==3.13.1
fsspec==2024.2.0
google-api-core==2.20.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
idna==3.10
isodate==0.6.1
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
packaging==24.1
pillow==10.2.0
pip==24.2
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydash==8.0.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
strictyaml==1.7.3
sympy==1.12
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.66.5
triton==2.2.0
typing_extensions==4.9.0
urllib3==2.2.3
wheel==0.44.0
